Remove testiii18naware.py because it didn't have any actual tests and was only used in one module, testi18nawareobject.py. This improves coverage.
